Introduces pixel alignment

+++ Pixl/Sources/PixlUI/Layout Types/EdgeInsets.swift
@@ -2,12 +2,12 @@ import Swift
 
 /// The inset distances for the sides of a rectangle.
 @frozen public struct EdgeInsets: Equatable, Sendable {
-    public var top: Double = 0
-    public var leading: Double = 0
-    public var bottom: Double = 0
-    public var trailing: Double = 0
+    public var top: Float = 0
+    public var leading: Float = 0
+    public var bottom: Float = 0
+    public var trailing: Float = 0
 
-    @inlinable public init(top: Double, leading: Double, bottom: Double, trailing: Double) {
+    @inlinable public init(top: Float, leading: Float, bottom: Float, trailing: Float) {
         self.top = top
         self.leading = leading
         self.bottom = bottom
